+++ Controladores/actualizacionesproveedor.php
@@ -4,6 +4,10 @@ require '../conexion.php';
 // Incluir la librería PHP QR Code
 include('../phpqrcode/qrlib.php');
 
+session_start();
+$rol = $_SESSION['rol']; // Obtener el rol del usuario
+
+
 // Búsqueda de proveedor
 $searchTerm = '';
 if (isset($_POST['search'])) {
@@ -11,13 +15,14 @@ if (isset($_POST['search'])) {
 }
 
 // Actualizar proveedor
-if (isset($_POST['actualizar'])) {
+if ($rol === 'admin' && isset($_POST['actualizar'])) {
     $id_ = $_POST['id'];
     $nombre_apellido_ = $_POST['nombre_apellido'];
     $proveedor_ = $_POST['proveedor'];
     $placas_vehiculos_ = $_POST['placas_vehiculos'];
     $modelo_marca_ = $_POST['modelo_marca'];
     $color_vehiculo_ = $_POST['color_vehiculo'];
+    $fecha_expiracion_ = $_POST['fecha_expiracion'];
 
     // Verificar si la placa ya existe para otro proveedor
     $sql = "SELECT id FROM proveedores WHERE placas_vehiculos = ? AND id != ?";
@@ -51,7 +56,7 @@ if (isset($_POST['actualizar'])) {
         }
 
         // Generar el contenido del QR usando los datos actualizados
-        $qrContent = "proveedor $proveedor_ \nNombre: $nombre_apellido_ \nPlacas: $placas_vehiculos_ \nVehículo: $modelo_marca_ ($color_vehiculo_)";
+        $qrContent = "proveedor|$proveedor_|\nNombre:$nombre_apellido_|\nPlacas:|$placas_vehiculos_|\nVehículo:$modelo_marca_|\n$color_vehiculo_";
 
         // Actualizar el contenido del QR (sobrescribir el archivo QR existente)
         QRcode::png($qrContent, $newQrFilePath);
@@ -59,10 +64,10 @@ if (isset($_POST['actualizar'])) {
         // Guardar la ruta o el nombre del archivo QR en la base de datos
         $qr_code_ = $newQrFilePath;
 
-        $sql = "UPDATE proveedores SET nombre_apellido = ?, proveedor = ?, placas_vehiculos = ?, modelo_marca = ?, color_vehiculo = ?, qr_code = ? WHERE id = ?";
+        $sql = "UPDATE proveedores SET nombre_apellido = ?, proveedor = ?, placas_vehiculos = ?, modelo_marca = ?, color_vehiculo = ?, fecha_expiracion = ?, qr_code = ? WHERE id = ?";
         $stmt = $conn->prepare($sql);
         
-        if ($stmt->execute([$nombre_apellido_, $proveedor_, $placas_vehiculos_, $modelo_marca_, $color_vehiculo_, $qr_code_, $id_])) {
+        if ($stmt->execute([$nombre_apellido_, $proveedor_, $placas_vehiculos_, $modelo_marca_, $color_vehiculo_, $fecha_expiracion_, $qr_code_, $id_])) {
             echo "<div class='alert alert-success' role='alert'>Proveedor y código QR actualizados correctamente</div>";
         } else {
             echo "<div class='alert alert-danger' role='alert'>Error al actualizar el proveedor</div>";
@@ -72,7 +77,7 @@ if (isset($_POST['actualizar'])) {
 
 
 // Eliminar proveedor
-if (isset($_GET['eliminar'])) {
+if (isset($_GET['eliminar']) && $rol === 'admin') {
     $id = $_GET['eliminar'];
 
     $sql = "DELETE FROM proveedores WHERE id = ?";
@@ -86,7 +91,7 @@ if (isset($_GET['eliminar'])) {
 }
 
 // Obtener proveedores
-$sql = "SELECT id, nombre_apellido, proveedor, placas_vehiculos, modelo_marca, color_vehiculo, qr_code, estado FROM proveedores";
+$sql = "SELECT id, nombre_apellido, proveedor, placas_vehiculos, modelo_marca, color_vehiculo, fecha_expiracion, qr_code, estado FROM proveedores";
 if (!empty($searchTerm)) {
     $sql .= " WHERE proveedor LIKE ?";
     $stmt = $conn->prepare($sql);
@@ -97,7 +102,7 @@ if (!empty($searchTerm)) {
 $proveedores = $stmt->fetchAll(PDO::FETCH_ASSOC);
 
 // Habilitar/Deshabilitar proveedor
-if (isset($_GET['habilitar']) || isset($_GET['deshabilitar'])) {
+if (($rol === 'admin') && (isset($_GET['habilitar']) || isset($_GET['deshabilitar']))) {
     $id = isset($_GET['habilitar']) ? $_GET['habilitar'] : $_GET['deshabilitar'];
     $estado = isset($_GET['habilitar']) ? 1 : 0; // 1 para habilitar, 0 para deshabilitar
 
@@ -171,6 +176,7 @@ if (isset($_GET['habilitar']) || isset($_GET['deshabilitar'])) {
         <div class="d-flex justify-content-end mb-2">
         <!-- Enlace a la derecha con margen superior y mejor formato -->
 
+
     <a href="../fpdf/reporteProveedor.php" target="_blank" class="btn btn-primary d-flex align-items-center ms-3 mt-3">
         <i class="bi bi-file-earmark-pdf-fill me-2"></i> Generar Reporte
     </a>
@@ -208,7 +214,8 @@ if (isset($_GET['habilitar']) || isset($_GET['deshabilitar'])) {
                     <th>Proveedor</th>
                     <th>Placas del Vehículo</th>
                     <th>Modelo y Marca</th>
-                    <th>Color del Vehículo</th>
+                    <th>Color</th>
+                    <th>Fecha Expiración</th>
                     <th>QR</th>
                     <th>Estado</th>
                     <th>Acciones</th>
@@ -223,11 +230,13 @@ if (isset($_GET['habilitar']) || isset($_GET['deshabilitar'])) {
                     <td><?php echo $row['placas_vehiculos']; ?></td>
                     <td><?php echo $row['modelo_marca']; ?></td>
                     <td><?php echo $row['color_vehiculo']; ?></td>
+                    <td><?php echo $row['fecha_expiracion']; ?></td>
                     <td>
-                <a href="#" data-bs-toggle="modal" data-bs-target="#modalQr" onclick="showQr('<?php echo $row['qr_code']; ?>', '<?php echo $row['nombre_apellido']; ?>', '<?php echo $row['proveedor']; ?>')">
+                <a href="#" data-bs-toggle="modal" data-bs-target="#modalQr" onclick="showQr('<?php echo $row['qr_code']; ?>', '<?php echo $row['nombre_apellido']; ?>', '<?php echo $row['proveedor']; ?>', '<?php echo $row['fecha_expiracion']; ?>')">
                     <img src="<?php echo $row['qr_code']; ?>" alt="QR Code" width="50">
                 </a>
             </td>
+            <?php if ($rol === 'admin'): ?>
                     <td><?php echo $row['estado'] ? 'Habilitado' : 'Deshabilitado'; ?></td>
                     <td>
                         <a href="actualizacionesproveedor.php?id=<?php echo $row['id']; ?>" class="btn btn-warning btn-sm">
@@ -245,6 +254,9 @@ if (isset($_GET['habilitar']) || isset($_GET['deshabilitar'])) {
                                 <i class="bi bi-toggle-on"></i>
                             </a>
                         <?php endif; ?>
+                        <?php else: ?>
+                
+                <?php endif; ?>
                     </td>
                 </tr>
                 <?php endforeach; ?>
@@ -260,14 +272,15 @@ if (isset($_GET['habilitar']) || isset($_GET['deshabilitar'])) {
         </div>
 
         <script>
-    function showQr(qrImage, nombre, proveedor) {
+    function showQr(qrImage, nombre, proveedor, fecha_expiracion) {
         // Establecer el QR de gran tamaño en el modal
         document.getElementById("qrImage").src = qrImage;
 
         // Mostrar la información correspondiente
         document.getElementById("qrInfo").innerHTML = `
             <strong>Nombre:</strong> ${nombre}<br>
-            <strong>Proveedor:</strong> ${proveedor}
+            <strong>Proveedor:</strong> ${proveedor}<br>
+            <strong>Fecha:</strong> ${fecha_expiracion}
         `;
     }
 </script>
